musb: sunxi: Fix support for H3/H5 OTG controller
The OTG controller on the H3/H5, like some of the previous chips,
omits the hardware feature register. Return a hardcoded value in
the I/O callback when this register offset is accessed.

Also, the H3/H5 controller supports less endpoints then on the
other SoCs. Set the proper value in the config structure.

Fixes: 3a61b080acee ("musb: sunxi: switch to the device model")
Signed-off-by: Chen-Yu Tsai <wens@csie.org>