sunxi: Enable USB OTG in host mode for Libre Computer Board ALL-H3-CC H3 ver.
The Libre Computer Board ALL-H3-CC has USB0, the USB OTG controller pair,
wired to a USB host port.

Enable the relevant controllers in the device tree, thus matching what
is in the device tree of the Linux kernel. Since it is host mode only,
we don't need to enable the MUSB driver, as the USB EHCI/OHCI hosts
can handle it.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>